Fix api,category,product

+++ app/src/main/java/com/onur/fastproudsearch/Product.java
@@ -13,7 +13,7 @@ public class Product {
     private String title;
     @SerializedName("price")
     @Expose
-    private Integer price;
+    private Double price;
     @SerializedName("description")
     @Expose
     private String description;
@@ -34,9 +34,7 @@ public class Product {
         return id;
     }
 
-    public void setId(Integer id) {
-        this.id = id;
-    }
+    public void setId(Integer id) {this.id = id;}
 
     public String getTitle() {
         return title;
@@ -46,11 +44,11 @@ public class Product {
         this.title = title;
     }
 
-    public Integer getPrice() {
+    public Double getPrice() {
         return price;
     }
 
-    public void setPrice(Integer price) {
+    public void setPrice(Double price) {
         this.price = price;
     }
